RenderAudio() working

+++ resid/resid-c-wrapper.cpp
@@ -105,10 +105,10 @@ bool ReSIDDmpPlayer_fillAudioBuffer(ReSIDDmpPlayer *dmpply)
     return dmpply->FillAudioBuffer();
 }
 
-unsigned long ReSIDDmpPlayer_RenderAudio(ReSIDDmpPlayer *dmpply,
-                                         unsigned int start_step,
-                                         unsigned int num_steps,
-                                         unsigned int buf_size, short *buffer)
+unsigned int ReSIDDmpPlayer_RenderAudio(ReSIDDmpPlayer *dmpply,
+                                        unsigned int start_step,
+                                        unsigned int num_steps,
+                                        unsigned int buf_size, short *buffer)
 {
     return dmpply->RenderAudio(start_step, num_steps, buf_size, buffer);
 }
